+++ khoros/objects/users.py
@@ -6,7 +6,7 @@
 :Example:           ``khoros.users.create(username='john_doe', email='john.doe@example.com')``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     29 Mar 2021
+:Modified Date:     07 Apr 2021
 """
 
 import warnings
@@ -95,8 +95,11 @@ def impersonate_user(khoros_object, user_login):
 
 
 def create(khoros_object, user_settings=None, login=None, email=None, password=None, first_name=None, last_name=None,
-           biography=None, sso_id=None, web_page_url=None, cover_image=None):
+           biography=None, sso_id=None, web_page_url=None, cover_image=None, ignore_exceptions=False):
     """This function creates a new user in the Khoros Community environment.
+
+    .. versionchanged:: 4.0.0
+       This function now returns the API response and the ``ignore_exceptions`` parameter has been introduced.
 
     .. versionchanged:: 3.3.0
        Updated ``khoros_object._settings`` to be ``khoros_object.core_settings``.
@@ -126,18 +129,21 @@ def create(khoros_object, user_settings=None, login=None, email=None, password=N
     :type web_page_url: str, None
     :param cover_image: The cover image to be used on the user's profile
     :type cover_image: str, None
-    :returns: None
-    :raises: :py:exc:`khoros.errors.exceptions.UserCreationError`
+    :param ignore_exceptions: Defines whether to raise the :py:exc:`khoros.errors.exceptions.UserCreationError`
+                              exception if the creation attempt fails (``False`` by default)
+    :type ignore_exceptions: bool
+    :returns: The response to the user creation API request
+    :raises: :py:exc:`KeyError`, :py:exc:`khoros.errors.exceptions.UserCreationError`
     """
     # TODO: Add functionality for followers, following, rank, roles, user_avatar and user_badges
     payload = structure_payload(user_settings, login, email, password, first_name, last_name, biography, sso_id,
                                 web_page_url, cover_image)
-    query_url = f"{khoros_object.core_settings['v2_base']}/users"
+    query_url = f"{khoros_object.core_settings.get('v2_base')}/users"
     headers = {'content-type': 'application/json'}
     response = api.post_request_with_retries(query_url, payload, auth_dict=khoros_object.auth, headers=headers)
-    if not api.query_successful(response):
-        raise errors.exceptions.UserCreationError(user=payload['login'], exc_msg=response['message'])
-    return
+    if not api.query_successful(response) and not ignore_exceptions:
+        raise errors.exceptions.UserCreationError(user=payload.get('login'), exc_msg=response.get('message'))
+    return response
 
 
 def process_user_settings(user_settings=None, user_id=None, albums=None, avatar=None, banned=None, biography=None,
@@ -285,7 +291,7 @@ def process_user_settings(user_settings=None, user_id=None, albums=None, avatar=
 
     # Overwrite any settings where fields are explicitly passed as arguments
     for field_name, field_value in default_settings.items():
-        if default_settings[field_name]:
+        if default_settings.get(field_name):
             user_settings[field_name] = field_value
 
     # Ensure the User ID uses 'id' rather than 'user_id' as the field name
@@ -298,6 +304,10 @@ def process_user_settings(user_settings=None, user_id=None, albums=None, avatar=
 def structure_payload(user_settings=None, login=None, email=None, password=None, first_name=None, last_name=None,
                       biography=None, sso_id=None, web_page_url=None, cover_image=None):
     """This function properly structures the payload to be passed when creating or manipulating users via the API.
+
+    .. versionchanged:: 4.0.0
+       Fixed an issue that was resulting in a :py:exc:`KeyError` exception potentially getting raised, and added
+       the missing ``type`` key that was preventing users from getting created successfully.
 
     :param user_settings: Allows all user settings to be passed to the function within a single dictionary
     :type user_settings: dict, None
@@ -322,6 +332,7 @@ def structure_payload(user_settings=None, login=None, email=None, password=None,
     :returns: The properly formatted payload within a dictionary
     """
     payload_mapping = {
+        'type': 'user',
         'biography': biography,
         'cover_image': cover_image,
         'email': email,
@@ -336,7 +347,7 @@ def structure_payload(user_settings=None, login=None, email=None, password=None,
     if user_settings:
         payload.update(user_settings)
     for field_name, field_value in payload_mapping.items():
-        if payload_mapping[field_name]:
+        if payload_mapping.get(field_name):
             payload[field_name] = field_value
     payload = {'data': payload}
     return payload
@@ -355,11 +366,20 @@ def delete(khoros_object, user_id, return_json=False):
     :param return_json: Determines if the API response should be returned in JSON format (``False`` by default)
     :type return_json: bool
     :returns: The API response (optionally in JSON format)
+    :raises: :py:exc:`khoros.errors.exceptions.FeatureNotConfiguredError`
     """
     # TODO: Allow other identifiers (e.g. login, email, etc.) to be provided instead of just the User ID
-    # TODO: Add a confirmation prompt that can be optionally disabled
     query_url = f"{khoros_object.core_settings['v2_base']}/users/{user_id}"
-    return api.delete(query_url, return_json, auth_dict=khoros_object.auth)
+    response = api.delete(query_url, return_json, auth_dict=khoros_object.auth)
+    if response.status_code == 403 and 'Feature is not configured' in response.text:
+        try:
+            identifier = response.text.split('identifier: ')[1].split('"')[0]
+            raise errors.exceptions.FeatureNotConfiguredError(identifier=identifier)
+        except IndexError:
+            raise errors.exceptions.FeatureNotConfiguredError()
+    if return_json:
+        response = response.json()
+    return response
 
 
 def _get_where_clause_for_user_id(_user_settings):